Create README.md
The fun and easy project “Guess the Number” is a short Java project that allows the user to guess the number generated by the computer & involves the following steps: The system generates a random number from a given range, say 1 to 100. The user is prompted to enter their given number in a displayed dialogue box. The computer then tells if the entered number matches the guesses number or it is higher/lower than the generated number. The game continues under the user guessing the number.